pyproject.toml
[tool.poetry]
name = "cityseg-project"
version = "0.1.0"
description = ""
authors = ["Aurélien Gruzon <aureliengruzon@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
tensorflow = "2.20.0"
numpy = ">=1.26,<2.1"
pandas = "^2.2"
pillow = "^10.0"
matplotlib = "^3.8"
opencv-python = "^4.9"
tqdm = "^4.66"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.27"}
python-multipart = "^0.0.9"
requests = "^2.31"
streamlit = "^1.31"
cityscapesscripts = "^2.2.4"
mlflow = "^3.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "jupyter (>=1.1.1,<2.0.0)"
]
